x86: mce: Avoid unnecessary CMCIs on CPUs that don't process them
CMCI is always broadcast to all CPUs on each corrected error.
When a CPU owns no bank this interrupt is wasted because the
interrupt handler will do nothing.

This patch masks the vector when a CPU doesn't own any banks,
this way the interrupt will never be raised. This saves
significant cycles on errors for hyper threaded systems,
because only one thread will have to process the CMCI.

Signed-off-by: Andi Kleen <ak@linux.intel.com>