x86: mce: Fix CPU offlining of shared MCE banks
When offlining a CPU the machine check bank shouldn't
be disabled until the last sharer of the bank is offlined.

Currently the code always clears all banks on offlining,
effectively disabling machine checks when any CPU on
a core goes down.

Discovering which banks are shared is unfortunately a bit
tricky.

This patch implements two heuristics to detect them:

- Assume that thread siblings always share all machine check banks
- When doing CMCI bank ownership discovery and a bank was already
owned by someone else and this is the first sibling to come alive
assume the bank is shared over the socket. That's a common situation
on Intel CPUs when the bank is implemented in the uncore.

Then it uses these heuristics to only clear machine checks
when the last sharer goes down.

Signed-off-by: Andi Kleen <ak@linux.intel.com>