x86: mce: Only round jiffies in poll timer when there are no errors
When there are errors and the error poll timer is in variable
frequency mode better be exact, power saving is not that important
in this case.
This avoids the problem that if the timer slows down
on less errors it would always jump to a second very quickly
because the jiffies rounding would round it up to a full
second.  But the slowdown is supposed to be smooth, not
contain large jumps, to get the best sampling of the errors.

It's better to slow down a little smoother, even
if it costs some power in the error case.

Signed-off-by: Andi Kleen <ak@linux.intel.com>