x86: mce: Convert /dev/mcelog to per CPU kfifo
Convert the old custom ring buffer to queue events
from the MCE handlers to user space to per CPU kfifo. Now
that kfifo is NMI safe this can be done.

If a new tracing mechanism comes around it could be used
too of course, but for compatibility the old one still
has to be supported anyways for old user space. So it makes
sense to implement it as nicely as possible.

This should fix various long-standing bugs, including
a known (but hard to hit in practice) live-lock condition.

This has the advantage that it scales nicely with the number
of CPUs unlike the old buffer which was of a very limited size.
Currently it uses 2k log buffer per CPU.

We've seen some problems where the old buffer could be overwhelmed
easily, this is also fixed here.

I removed the old "anchor with signature" scheme. Now that
the log buffer is distributed over many buffers it's unlikely
to be useful anymore. Also I don't know of any users anyways.

Requires the earlier kfifo patch to add peek with offset.

Signed-off-by: Andi Kleen <ak@linux.intel.com>